requirements.txt
discord.py==1.3.1
praw
coloredlogs
pyyaml
aiosqlite3
matplotlib
